Update best_producer.c
best_producer: 
The program creates 3 producer threads and one consumer; moreover, it provides a dynamic allocated bidimensional vector of randomized int (dimension of matrix : n, passed as argument); finally, a stack with the following record type: id of a thread, an integer value (dimension of stack : n*n) . The main thread start the other ones: producers go on taking elements at random from the matrix and putting them on the stack (this was implemented guaranteeing mutual exclusion of writing for every thread); meanwhile the consumer wait on a condition variable (swapping out from RAM) and is signaled when the stack is fullfilled, then it prints all elements in the stack and the tid of the thread that put the highest number of values in the stack; finally every thread terminates. For synchronization two techniques were used: a named POSIX semaphore, initialized with 1 resource for producers and a condition variable (with an associated mutex) for consumer (in order to avoid busy waiting).